refactor: 코드 포맷팅 및 사용하지 않는 의존성 정리
