+++ src/main/java/ddip/me/ddipbe/global/config/session/SessionConfig.java
@@ -1,6 +1,5 @@
 package ddip.me.ddipbe.global.config.session;
 
-import ddip.me.ddipbe.global.config.session.SessionMemberIdResolver;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
